refactor(project): convertir ajustes de sheet a página dedicada
