+++ src/features/project/components/ProjectSidebarContent.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react"
 import { IconChartBar, IconLayoutKanban, IconLogout, IconSettings } from "@tabler/icons-react"
 import { Link, useLocation, useNavigate, useParams } from "react-router"
 import {
@@ -14,11 +13,11 @@ import {
 import { Button } from "@/shared"
 import { supabase } from "@/shared/supabase"
 import { useProjectsContext } from "../context/projectsCtx"
-import { ProjectSettingsSheet } from "./ProjectSettingsSheet"
 
 const PROJECT_VIEWS = [
   { label: "Tablero", icon: IconLayoutKanban, path: (id: string) => `/projects/${id}` },
   { label: "Analytics", icon: IconChartBar, path: (id: string) => `/projects/${id}/analytics` },
+  { label: "Ajustes", icon: IconSettings, path: (id: string) => `/projects/${id}/settings` },
 ]
 
 export function ProjectSidebarContent() {
@@ -27,7 +26,6 @@ export function ProjectSidebarContent() {
   const { id: activeId } = useParams()
   const location = useLocation()
   const navigate = useNavigate()
-  const [settingsOpen, setSettingsOpen] = useState(false)
 
   const handleLogout = async () => {
     await supabase.auth.signOut()
@@ -98,14 +96,6 @@ export function ProjectSidebarContent() {
                             </Link>
                           )
                         })}
-                        {/* Ajustes — abre Sheet, no navega */}
-                        <button
-                          onClick={() => setSettingsOpen(true)}
-                          className="flex items-center gap-2 pl-8 pr-3 py-1.5 rounded-md text-sm transition-colors text-muted-foreground hover:text-primary hover:bg-muted w-full text-left"
-                        >
-                          <IconSettings size={14} />
-                          Ajustes
-                        </button>
                       </div>
                     )}
                   </SidebarMenuItem>
@@ -143,14 +133,6 @@ export function ProjectSidebarContent() {
         </div>
       </SidebarFooter>
 
-      {/* Settings Sheet — solo se monta si hay proyecto activo */}
-      {activeId && (
-        <ProjectSettingsSheet
-          projectId={activeId}
-          open={settingsOpen}
-          onOpenChange={setSettingsOpen}
-        />
-      )}
     </>
   )
 }
