+++ src/features/project/index.ts
@@ -2,7 +2,7 @@ export { ProjectsPage } from "./components/ProjectsPage"
 export { ProjectCard } from "./components/ProjectCard"
 export { ProjectSidebarContent } from "./components/ProjectSidebarContent"
 export { CreateProjectModal } from "./components/CreateProjectModal"
-export { ProjectSettingsSheet } from "./components/ProjectSettingsSheet"
+export { ProjectSettingsPage } from "./components/ProjectSettingsPage"
 export { ProjectsProvider } from "./context/ProjectsContext"
 export { useProjectsContext } from "./context/projectsCtx"
 export { useProjects } from "./hooks/useProjects"
